First interactions with a Repository and displaying on the home page
