+++ TwoWeeksReady.Common/FamilyMembers/FamilyMember.cs
@@ -1,0 +1,10 @@
+using System;
+
+namespace TwoWeeksReady.Common.FamilyMembers
+{
+	public class FamilyMember
+	{
+		public Guid Id { get; set; }
+		public string Name { get; set; }
+	}
+}
