+++ admin/TwoWeeksReady.Admin/App.razor
