+++ admin/TwoWeeksReady.Admin/Data/IRepository.cs
@@ -1,0 +1,18 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+using TwoWeeksReady.Common.EmergencyKits;
+
+namespace TwoWeeksReady.Admin.Data
+{
+	public interface IRepository
+	{
+
+		Task<BaseKit> GetBaseKitById(string id);
+
+		Task<IEnumerable<BaseKit>> GetAllBaseKits();
+
+	}
+
+}
