+++ admin/TwoWeeksReady.Admin/Pages/Error.cshtml
@@ -1,5 +1,5 @@
 ﻿@page
-@model adminweb.Pages.ErrorModel
+@model TwoWeeksReady.Admin.Pages.ErrorModel
 
 <!DOCTYPE html>
 <html>
