+++ admin/TwoWeeksReady.Admin/Pages/Index.razor
@@ -1,0 +1,44 @@
+@page "/"
+@inject IRepository Repository
+
+<h1>Administer the Base Kits for Two Weeks Ready</h1>
+
+<h3>Current Basekits Defined:</h3>
+
+@if (BaseKits != null && BaseKits.Any()) {
+
+<table>
+	<thead>
+	 <tr>
+		<th>Name</th>
+		<th>Line Item Count</th>
+	</tr>
+ </thead>
+ <tbody>
+	 @foreach	(var kit in BaseKits) {
+		
+		 <tr>
+			<td><a href="/Kits/@kit.Id"><i class="mdi @kit.Icon"
+               aria-hidden="true"></i>  @kit.Name</a></td>
+			<td>@kit.Items.Count()</td>
+		</tr>
+			 }
+ </tbody>
+
+</table>
+
+} else {
+	<p>Loading...</p>
+}
+
+@code {
+
+	public IEnumerable<BaseKit> BaseKits { get; set; }
+
+	protected override async Task OnInitializedAsync() {
+
+			BaseKits = await Repository.GetAllBaseKits();
+
+	}
+
+}
