+++ admin/TwoWeeksReady.Admin/Pages/_Host.cshtml
@@ -1,5 +1,5 @@
 ﻿@page "/"
-@namespace adminweb.Pages
+@namespace TwoWeeksReady.Admin.Pages
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
 @{
     Layout = null;
@@ -10,11 +10,12 @@
 <head>
     <meta charset="utf-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>adminweb</title>
+    <title>Two Weeks Ready :: Application Administration</title>
     <base href="~/" />
     <link rel="stylesheet" href="css/bootstrap/bootstrap.min.css" />
+    <link rel="stylesheet" href="//cdn.materialdesignicons.com/5.4.55/css/materialdesignicons.min.css" />
     <link href="css/site.css" rel="stylesheet" />
-    <link href="adminweb.styles.css" rel="stylesheet" />
+    <link href="TwoWeeksReady.Admin.styles.css" rel="stylesheet" />
 </head>
 <body>
     <component type="typeof(App)" render-mode="ServerPrerendered" />
