+++ admin/TwoWeeksReady.Admin/TwoWeeksReady.Admin.csproj
@@ -5,7 +5,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <Folder Include="Data\" />
+    <ProjectReference Include="..\..\TwoWeeksReady.Common\TwoWeeksReady.Common.csproj" />
   </ItemGroup>
 
 </Project>
