+++ admin/TwoWeeksReady.Admin/_Imports.razor
@@ -6,5 +6,8 @@
 @using Microsoft.AspNetCore.Components.Web
 @using Microsoft.AspNetCore.Components.Web.Virtualization
 @using Microsoft.JSInterop
-@using adminweb
-@using adminweb.Shared
+@using TwoWeeksReady.Admin
+@using TwoWeeksReady.Admin.Data
+@using TwoWeeksReady.Admin.Shared
+@using TwoWeeksReady.Common
+@using TwoWeeksReady.Common.EmergencyKits
