+++ admin/TwoWeeksReady.Admin/wwwroot/css/open-iconic/FONT-LICENSE
