+++ admin/TwoWeeksReady.Admin/wwwroot/css/open-iconic/README.md
