+++ admin/TwoWeeksReady.Admin/wwwroot/css/open-iconic/font/fonts/open-iconic.woff
