+++ admin/admin.sln
@@ -1,9 +1,13 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.26124.0
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.31004.235
 MinimumVisualStudioVersion = 15.0.26124.0
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "adminweb", "adminweb\adminweb.csproj", "{D69A7662-03D8-4FDE-87B6-D732DE3396DF}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TwoWeeksReady", "..\api\TwoWeeksReady\TwoWeeksReady.csproj", "{E387C361-5643-47DA-B395-9FA72AD21E67}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TwoWeeksReady.Common", "..\TwoWeeksReady.Common\TwoWeeksReady.Common.csproj", "{484A0768-D116-43FF-8476-F96F8BFB3875}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TwoWeeksReady.Admin", "TwoWeeksReady.Admin\TwoWeeksReady.Admin.csproj", "{428E5852-1B79-4EBC-B48F-BAB47BBD2C5A}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -14,21 +18,48 @@ Global
 		Release|x64 = Release|x64
 		Release|x86 = Release|x86
 	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{E387C361-5643-47DA-B395-9FA72AD21E67}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E387C361-5643-47DA-B395-9FA72AD21E67}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E387C361-5643-47DA-B395-9FA72AD21E67}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{E387C361-5643-47DA-B395-9FA72AD21E67}.Debug|x64.Build.0 = Debug|Any CPU
+		{E387C361-5643-47DA-B395-9FA72AD21E67}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{E387C361-5643-47DA-B395-9FA72AD21E67}.Debug|x86.Build.0 = Debug|Any CPU
+		{E387C361-5643-47DA-B395-9FA72AD21E67}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E387C361-5643-47DA-B395-9FA72AD21E67}.Release|Any CPU.Build.0 = Release|Any CPU
+		{E387C361-5643-47DA-B395-9FA72AD21E67}.Release|x64.ActiveCfg = Release|Any CPU
+		{E387C361-5643-47DA-B395-9FA72AD21E67}.Release|x64.Build.0 = Release|Any CPU
+		{E387C361-5643-47DA-B395-9FA72AD21E67}.Release|x86.ActiveCfg = Release|Any CPU
+		{E387C361-5643-47DA-B395-9FA72AD21E67}.Release|x86.Build.0 = Release|Any CPU
+		{484A0768-D116-43FF-8476-F96F8BFB3875}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{484A0768-D116-43FF-8476-F96F8BFB3875}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{484A0768-D116-43FF-8476-F96F8BFB3875}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{484A0768-D116-43FF-8476-F96F8BFB3875}.Debug|x64.Build.0 = Debug|Any CPU
+		{484A0768-D116-43FF-8476-F96F8BFB3875}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{484A0768-D116-43FF-8476-F96F8BFB3875}.Debug|x86.Build.0 = Debug|Any CPU
+		{484A0768-D116-43FF-8476-F96F8BFB3875}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{484A0768-D116-43FF-8476-F96F8BFB3875}.Release|Any CPU.Build.0 = Release|Any CPU
+		{484A0768-D116-43FF-8476-F96F8BFB3875}.Release|x64.ActiveCfg = Release|Any CPU
+		{484A0768-D116-43FF-8476-F96F8BFB3875}.Release|x64.Build.0 = Release|Any CPU
+		{484A0768-D116-43FF-8476-F96F8BFB3875}.Release|x86.ActiveCfg = Release|Any CPU
+		{484A0768-D116-43FF-8476-F96F8BFB3875}.Release|x86.Build.0 = Release|Any CPU
+		{428E5852-1B79-4EBC-B48F-BAB47BBD2C5A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{428E5852-1B79-4EBC-B48F-BAB47BBD2C5A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{428E5852-1B79-4EBC-B48F-BAB47BBD2C5A}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{428E5852-1B79-4EBC-B48F-BAB47BBD2C5A}.Debug|x64.Build.0 = Debug|Any CPU
+		{428E5852-1B79-4EBC-B48F-BAB47BBD2C5A}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{428E5852-1B79-4EBC-B48F-BAB47BBD2C5A}.Debug|x86.Build.0 = Debug|Any CPU
+		{428E5852-1B79-4EBC-B48F-BAB47BBD2C5A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{428E5852-1B79-4EBC-B48F-BAB47BBD2C5A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{428E5852-1B79-4EBC-B48F-BAB47BBD2C5A}.Release|x64.ActiveCfg = Release|Any CPU
+		{428E5852-1B79-4EBC-B48F-BAB47BBD2C5A}.Release|x64.Build.0 = Release|Any CPU
+		{428E5852-1B79-4EBC-B48F-BAB47BBD2C5A}.Release|x86.ActiveCfg = Release|Any CPU
+		{428E5852-1B79-4EBC-B48F-BAB47BBD2C5A}.Release|x86.Build.0 = Release|Any CPU
+	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{D69A7662-03D8-4FDE-87B6-D732DE3396DF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D69A7662-03D8-4FDE-87B6-D732DE3396DF}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D69A7662-03D8-4FDE-87B6-D732DE3396DF}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{D69A7662-03D8-4FDE-87B6-D732DE3396DF}.Debug|x64.Build.0 = Debug|Any CPU
-		{D69A7662-03D8-4FDE-87B6-D732DE3396DF}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{D69A7662-03D8-4FDE-87B6-D732DE3396DF}.Debug|x86.Build.0 = Debug|Any CPU
-		{D69A7662-03D8-4FDE-87B6-D732DE3396DF}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D69A7662-03D8-4FDE-87B6-D732DE3396DF}.Release|Any CPU.Build.0 = Release|Any CPU
-		{D69A7662-03D8-4FDE-87B6-D732DE3396DF}.Release|x64.ActiveCfg = Release|Any CPU
-		{D69A7662-03D8-4FDE-87B6-D732DE3396DF}.Release|x64.Build.0 = Release|Any CPU
-		{D69A7662-03D8-4FDE-87B6-D732DE3396DF}.Release|x86.ActiveCfg = Release|Any CPU
-		{D69A7662-03D8-4FDE-87B6-D732DE3396DF}.Release|x86.Build.0 = Release|Any CPU
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {172BCC3D-739F-4C3D-8415-C1D0C1E8507C}
 	EndGlobalSection
 EndGlobal
