--- admin/adminweb/Pages/Index.razor
@@ -1,7 +1,0 @@
-@page "/"
-
-<h1>Hello, world!</h1>
-
-Welcome to your new app.
-
-<SurveyPrompt Title="How is Blazor working for you?" />
