+++ api/TwoWeeksReady/EmergencyKits/BaseKitsApi.cs
@@ -1,4 +1,4 @@
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
@@ -14,47 +14,48 @@ using Microsoft.Azure.WebJobs.Host;
 using Microsoft.Extensions.Logging;
 using Newtonsoft.Json;
 using AzureFunctions.OidcAuthentication;
+using TwoWeeksReady.Common.EmergencyKits;
 
 namespace TwoWeeksReady.EmergencyKits
 {
-    public class BaseKitsApi
+  public class BaseKitsApi
+  {
+    private readonly IApiAuthentication _apiAuthentication;
+
+    public BaseKitsApi(IApiAuthentication apiAuthentication)
     {
-        private readonly IApiAuthentication _apiAuthentication;
-
-        public BaseKitsApi(IApiAuthentication apiAuthentication)
-        {
-            _apiAuthentication = apiAuthentication;
-        }
-
-        [FunctionName("basekits")]
-        public async Task<IActionResult> GetKits(
-            [HttpTrigger(AuthorizationLevel.Anonymous, "get", Route = null)]
-            HttpRequest req,
-            [CosmosDB( databaseName: "2wr", collectionName: "basekits", ConnectionStringSetting = "CosmosDBConnection")]
-            DocumentClient client,
-            ILogger log)
-        {
-          
-            log.LogInformation($"Getting list of base kits");      
-            var authorizationResult = await _apiAuthentication.AuthenticateAsync(req.Headers);
-            if (authorizationResult.Failed)
-            {
-                log.LogWarning(authorizationResult.FailureReason);
-                return new UnauthorizedResult();
-            }      
-            Uri collectionUri = UriFactory.CreateDocumentCollectionUri("2wr", "basekits");
-            var query = client.CreateDocumentQuery<BaseKit>(collectionUri, new FeedOptions{EnableCrossPartitionQuery = true})
-                                .AsDocumentQuery();
-
-            var baseKits = new List<BaseKit>();
-            while (query.HasMoreResults)
-            {
-                var result = await query.ExecuteNextAsync<BaseKit>();
-                baseKits.AddRange(result);            
-            }
-
-            return new OkObjectResult(baseKits);
-        }
-
+      _apiAuthentication = apiAuthentication;
     }
+
+    [FunctionName("basekits")]
+    public async Task<IActionResult> GetKits(
+        [HttpTrigger(AuthorizationLevel.Anonymous, "get", Route = null)]
+            HttpRequest req,
+        [CosmosDB( databaseName: "2wr", collectionName: "basekits", ConnectionStringSetting = "CosmosDBConnection")]
+            DocumentClient client,
+        ILogger log)
+    {
+
+      log.LogInformation($"Getting list of base kits");
+      var authorizationResult = await _apiAuthentication.AuthenticateAsync(req.Headers);
+      if (authorizationResult.Failed)
+      {
+        log.LogWarning(authorizationResult.FailureReason);
+        return new UnauthorizedResult();
+      }
+      Uri collectionUri = UriFactory.CreateDocumentCollectionUri("2wr", "basekits");
+      var query = client.CreateDocumentQuery<BaseKit>(collectionUri, new FeedOptions { EnableCrossPartitionQuery = true })
+                          .AsDocumentQuery();
+
+      var baseKits = new List<BaseKit>();
+      while (query.HasMoreResults)
+      {
+        var result = await query.ExecuteNextAsync<BaseKit>();
+        baseKits.AddRange(result);
+      }
+
+      return new OkObjectResult(baseKits);
+    }
+
+  }
 }
