--- api/TwoWeeksReady/FamilyMembers/FamilyMember.cs
@@ -1,10 +1,0 @@
-using System;
-
-namespace TwoWeeksReady.FamilyMembers
-{
-    public class FamilyMember 
-    {
-        public Guid Id {get; set;}
-        public string Name {get; set;}
-    }
-}
