+++ api/TwoWeeksReady/FamilyMembers/FamilyMembersApi.cs
@@ -10,6 +10,7 @@ using Microsoft.Azure.WebJobs;
 using Microsoft.Azure.WebJobs.Extensions.Http;
 using Microsoft.Azure.WebJobs.Host;
 using Microsoft.Extensions.Logging;
+using TwoWeeksReady.Common.FamilyMembers;
 
 namespace TwoWeeksReady.FamilyMembers
 {
